+++ ng-metadata.ts
@@ -1,7 +1,7 @@
-export { Optional } from './src/di/optional';
-export { Host } from './src/di/host';
-export { Inject } from './src/di/inject';
-export { Injectable } from './src/di/injectable';
+export { Optional } from './src/optional';
+export { Host } from './src/host';
+export { Inject } from './src/inject';
+export { Injectable } from './src/injectable';
 export { Pipe } from './src/pipe';
 export {
   Component,
